Add repository package and service package.

+++ src/main/java/entity/Matches.java
@@ -5,12 +5,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class Matches {
-    private Integer id;
+    private long id;
     private Players player1;
     private Players player2;
     private Players winner;
