+++ src/main/java/entity/Players.java
@@ -5,12 +5,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 
 public class Players {
-    private Integer id;
+    private long id;
     private String name;
 }
